+++ src/backend/core/management/commands/clean_pending_files.py
@@ -6,7 +6,6 @@ from django.core.management.base import BaseCommand, CommandError
 from django.utils import timezone
 
 from core.models import File, FileUploadStateChoices
-from core.tasks.file import process_file_deletion
 
 
 class Command(BaseCommand):
@@ -37,11 +36,7 @@ class Command(BaseCommand):
 
         count = 0
         for file in files.iterator():
-            # This check shouldn't happen, but just in case we do it to avoid an error
-            if not file.deleted_at:
-                file.soft_delete()
             file.hard_delete()
-            process_file_deletion(file.id)
             count += 1
 
         self.stdout.write(f"Cleaned {count} stale pending file(s).")